+++ old-ui/app/components/add-token/add-token.container.js
@@ -1,13 +1,14 @@
 import { connect } from 'react-redux'
 import AddToken from './add-token.component'
 
-const { setPendingTokens, clearPendingTokens } = require('../../../actions')
+const { setPendingTokens, clearPendingTokens, displayWarning, goHome, addToken, showConfirmAddTokensPage } = require('../../../../ui/app/actions')
 
 const mapStateToProps = ({ metamask }) => {
-  const { identities, tokens, pendingTokens } = metamask
+  const { identities, tokens, pendingTokens, network } = metamask
   return {
     identities,
     tokens,
+    network,
     pendingTokens,
   }
 }
@@ -15,8 +16,13 @@ const mapStateToProps = ({ metamask }) => {
 const mapDispatchToProps = dispatch => {
   return {
     setPendingTokens: tokens => dispatch(setPendingTokens(tokens)),
+    showConfirmAddTokensPage: () => dispatch(showConfirmAddTokensPage()),
     clearPendingTokens: () => dispatch(clearPendingTokens()),
+    displayWarning: (warn) => dispatch(displayWarning(warn)),
+    goHome: () => dispatch(goHome()),
+    addToken: (address, symbol, decimals, token) => dispatch(addToken(address, symbol, decimals, token)),
   }
 }
+
 
 export default connect(mapStateToProps, mapDispatchToProps)(AddToken)
