+++ old-ui/app/components/confirm-add-token/confirm-add-token.container.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 import ConfirmAddToken from './confirm-add-token.component'
 
-const { addTokens, clearPendingTokens, goHome } = require('../../../../ui/app/actions')
+const { addTokens, clearPendingTokens, goHome, showAddTokenPage } = require('../../../../ui/app/actions')
 
 const mapStateToProps = ({ metamask }) => {
   const { pendingTokens } = metamask
@@ -15,6 +15,7 @@ const mapDispatchToProps = dispatch => {
     addTokens: tokens => dispatch(addTokens(tokens)),
     clearPendingTokens: () => dispatch(clearPendingTokens()),
     goHome: () => dispatch(goHome()),
+    showAddTokenPage: () => dispatch(showAddTokenPage()),
   }
 }
